+++ lib/src/clojure_lsp/feature/completion.clj
@@ -88,18 +88,6 @@
     :else
     :reference))
 
-(defn ^:private var-defs-in-ns-named [analysis ns-str name-syms]
-  (let [ns (symbol ns-str)
-        name-set (set name-syms)]
-    (into []
-          (comp (mapcat val)
-                (filter #(= ns (:ns %)))
-                ;; clojure.core is 98% var-definitions, so faster to filter by
-                ;; name-set before filtering for var defs.
-                (filter #(contains? name-set (:name %)))
-                (q/xf-var-defs false))
-          analysis)))
-
 (defn ^:private element->label [{:keys [alias bucket] :as element} cursor-alias priority]
   (cond
     (= :keywords bucket)
@@ -286,30 +274,30 @@
                            (string/starts-with? (:name %) name))))
          (mapv #(element->completion-item % alias :alias-keyword)))))
 
-(defn ^:private with-core-items [matches-fn analysis {:keys [filename ns-name symbols priority]}]
-  (let [matches (filter (comp matches-fn str) symbols)
-        elem-by-name (->> (var-defs-in-ns-named analysis ns-name matches)
-                          (medley/index-by :name))]
-    (map (fn [sym] {:label (str sym)
-                    :kind (element->completion-item-kind (elem-by-name sym))
-                    :data {"filename" filename
-                           "name" (str sym)
-                           "ns" ns-name}
-                    :detail (str ns-name "/" sym)
-                    :priority priority})
-         matches)))
+(defn ^:private with-core-items [matches-fn {:keys [filename ns-name symbols priority]}]
+  (keep (fn [{:keys [name kind]}]
+          (let [sym-name (str name)]
+            (when (matches-fn sym-name)
+              {:label sym-name
+               :kind kind
+               :data {"filename" filename
+                      "name" sym-name
+                      "ns" ns-name}
+               :detail (str ns-name "/" sym-name)
+               :priority priority})))
+        symbols))
 
-(defn ^:private with-clojure-core-items [matches-fn analysis]
-  (with-core-items matches-fn analysis {:filename "/clojure.core.clj"
-                                        :ns-name "clojure.core"
-                                        :symbols common-sym/core-syms
-                                        :priority :clojure-core}))
+(defn ^:private with-clojure-core-items [matches-fn]
+  (with-core-items matches-fn {:filename "/clojure.core.clj"
+                               :ns-name "clojure.core"
+                               :symbols common-sym/clj-syms
+                               :priority :clojure-core}))
 
-(defn ^:private with-clojurescript-items [matches-fn analysis]
-  (with-core-items matches-fn analysis {:filename "/cljs.core.cljs"
-                                        :ns-name "cljs.core"
-                                        :symbols common-sym/cljs-syms
-                                        :priority :clojurescript-core}))
+(defn ^:private with-clojurescript-items [matches-fn]
+  (with-core-items matches-fn {:filename "/cljs.core.cljs"
+                               :ns-name "cljs.core"
+                               :symbols common-sym/cljs-syms
+                               :priority :clojurescript-core}))
 
 (defn ^:private with-java-items [matches-fn]
   (concat
@@ -378,16 +366,7 @@
             analysis (:analysis db)
             current-ns-elements (get analysis filename)
             support-snippets? (get-in db [:client-capabilities :text-document :completion :completion-item :snippet-support] false)
-            other-ns-elements (into []
-                                    (comp
-                                      (q/filter-project-analysis-xf db)
-                                      (mapcat val))
-                                    (dissoc analysis filename))
-            external-ns-elements (into []
-                                       (comp
-                                         (q/filter-external-analysis-xf db)
-                                         (mapcat val))
-                                       (dissoc analysis filename))
+            all-other-ns-elements (mapcat val (dissoc analysis filename))
             cursor-element (q/find-element-under-cursor analysis filename row col)
             cursor-value (if (= :vector (z/tag cursor-loc))
                            ""
@@ -415,16 +394,16 @@
                               (contains? (q/find-all-ns-definition-names analysis) (symbol cursor-value-or-ns)))
             items (cond
                     inside-refer?
-                    (with-refer-elements matches-fn cursor-loc (concat other-ns-elements external-ns-elements))
+                    (with-refer-elements matches-fn cursor-loc all-other-ns-elements)
 
                     inside-require?
-                    (cond-> (with-ns-definition-elements matches-fn (concat other-ns-elements external-ns-elements))
+                    (cond-> (with-ns-definition-elements matches-fn all-other-ns-elements)
                       (and support-snippets?
                            simple-cursor?)
                       (merging-snippets cursor-loc next-loc matches-fn settings))
 
                     aliased-keyword-value?
-                    (with-elements-from-aliased-keyword cursor-loc cursor-element analysis filename (concat other-ns-elements external-ns-elements))
+                    (with-elements-from-aliased-keyword cursor-loc cursor-element analysis filename all-other-ns-elements)
 
                     :else
                     (cond-> []
@@ -438,11 +417,11 @@
                       (or simple-cursor?
                           keyword-value?)
                       (-> (into (with-element-items matches-fn uri cursor-element current-ns-elements row col))
-                          (into (with-clojure-core-items matches-fn analysis)))
+                          (into (with-clojure-core-items matches-fn)))
 
                       (and simple-cursor?
                            (supports-cljs? uri))
-                      (into (with-clojurescript-items matches-fn analysis))
+                      (into (with-clojurescript-items matches-fn))
 
                       (and simple-cursor?
                            (supports-clj-core? uri))